backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sentence-transformers>=2.2.2
torch>=2.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas